Pulled up half the functions of ConsoleHost into a shared context.
